Add first version of RegisterExerciseScreen

+++ shared/src/commonMain/kotlin/com/aitrainingapp/domain/model/TrainingSeries.kt
@@ -3,7 +3,7 @@ package com.aitrainingapp.domain.model
 data class TrainingSeries(
     val id: Int,
     val trainingType: String,
-    val weight: Double,
+    val weight: Float,
     val reps: Int,
     val sets: Int,
     val rpe: Int,
